Fix rendering `name` attribute, reorganize internal types
Previous work missed to properly set `name` on each form control (inputs) for proper form markup. This was not needed for functionality, but we should strive for proper markup.

Unfortunately this triggered also a cascade of type refactorings, as for internal Glint types to all pass, `name` (which maps to a key of the passed `@data`) must be a string, and not any other type that `keyof` can be (number or symbol). Now we are a bit stricter about types (and more verbose 🙈), dealing only with string-typed keys and such subsets of form data.